fix mount, remove old/unecessary registrations
