+++ packages/@ember/engine/instance.js
@@ -172,7 +172,7 @@ const EngineInstance = EmberObject.extend(RegistryProxyMixin, ContainerProxyMixi
   cloneParentDependencies() {
     let parent = getEngineParent(this);
 
-    let registrations = ['route:basic', 'service:-routing', 'service:-glimmer-environment'];
+    let registrations = ['route:basic', 'service:-routing'];
 
     registrations.forEach(key => this.register(key, parent.resolveRegistration(key)));
 
@@ -185,7 +185,6 @@ const EngineInstance = EmberObject.extend(RegistryProxyMixin, ContainerProxyMixi
       '-view-registry:main',
       `renderer:-${env.isInteractive ? 'dom' : 'inert'}`,
       'service:-document',
-      P`template-compiler:main`,
     ];
 
     if (env.isInteractive) {
